EnemyDetection is ready for testing
I have set up a basic system.
Vision indicators: This is made with a script and is represented by a white square.
Detection variable: It's in place and can be interacted with. (It still needs to be properly linked to the UI we're using. This is also an issue with the enemies.)
Different states of detection: The enemy changes color depending on the detection value. This will, of course, be changed to something else. But the basic system can be used by Justin to modify enemy movement.

Fake Breakdown: Press Q to become semi-transparent, which slows down detection increase. Semi-transparency needs to be replaced with another sprite. You cannot move while the ability is active.

The balancing of numbers will come later.